Added shift duration fix
My Schedule displays 8 hour shifts as 7.4 hours to account for 36 minute lunch break, but for some that break is paid and you are still here for this shift, so this fix swaps all shifts (up to 12 hours) to the correct number of hours you're there for / paid for. This is a temporary solution which ignores the count at the top, which is still wrong. Further updates here are needed, but that's for the future.

Also, set version number to 1.1